Move protocol tests under tests/protocols/
